Deprecate sk_types.h and move pscontext class to a separate file pscontext.h
