+++ src/runtime_src/core/edge/include/CMakeLists.txt
@@ -3,10 +3,11 @@
 #
 set(MPSOC_HEADER_SRC
   xclhal2_mpsoc.h
+  pscontext.h
   sk_types.h)
 
 install (FILES ${MPSOC_HEADER_SRC} DESTINATION ${XRT_INSTALL_INCLUDE_DIR})
 
 message("-- XRT header files for MPSoC only")
 message("-- xclhal2_mpsoc.h")
-message("-- sk_types.h")
+message("-- pscontext.h")
